Divided pages and imported inside Main Layout, floating button to reload random cards in homepage
