+++ lib/pages/collection_page.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class CollectionPage extends StatelessWidget {
+  const CollectionPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(Icons.collections, size: 64, color: Colors.grey),
+          SizedBox(height: 16),
+          Text(
+            'Collection',
+            style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+          ),
+          SizedBox(height: 8),
+          Text('Le tue carte personali'),
+        ],
+      ),
+    );
+  }
+}
