+++ lib/pages/watchlist_page.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class WatchlistPage extends StatelessWidget {
+  const WatchlistPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(Icons.favorite, size: 64, color: Colors.grey),
+          SizedBox(height: 16),
+          Text(
+            'Watchlist',
+            style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+          ),
+          SizedBox(height: 8),
+          Text('Carte che stai monitorando'),
+        ],
+      ),
+    );
+  }
+}
